Returned const types and split lab1.c to 3 files
split lab1.c  to task1.c task2.c and lab1.c
