+++ Lab1C/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(Lab1C lab1.c)
+add_executable(Lab1C lab1.c task1.c task2.c)
 target_include_directories(Lab1C PUBLIC ../LibraryC)
 target_link_libraries(Lab1C LibraryC)
