+++ LibraryC/array.c
@@ -44,7 +44,7 @@ void array_delete(Array *arr)
 }
 
 // returns specified array element
-Data array_get(Array *arr, size_t index)
+Data array_get(const Array *arr, size_t index)
 {
     if (index >= arr->size){
         exit(1);
@@ -65,7 +65,7 @@ void array_set(Array *arr, size_t index, Data value)
 }
 
 // returns array size
-size_t array_size( Array *arr)
+size_t array_size(const Array *arr)
 {
     if (!arr){
         exit(1);
