+++ LibraryC/array.h
@@ -23,13 +23,13 @@ Array *array_create(size_t size, FFree f);
 void array_delete(Array *arr);
 
 // returns specified array element
-Data array_get( Array *arr, size_t index);
+Data array_get(const Array *arr, size_t index);
 
 // sets the specified array element to the value
 void array_set(Array *arr, size_t index, Data value);
 
 // returns array size
-size_t array_size( Array *arr);
+size_t array_size(const Array *arr);
 
 #ifdef __cplusplus
 }
